flask/requirements.txt
gunicorn~=20.1.0
Flask==2.1.1
intersystems-irispython~=3.2.0
psycopg2-binary
Flask-SQLAlchemy
ddtrace

